+++ samples/tfm-comm/tfm-comm3/packet.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <stddef.h>
+#include <stdint.h>
 
 #include <mpi.h>
 
@@ -12,7 +13,7 @@ typedef struct {
     MPI_Request req; // MPI request for handling synchronization
 } Packet;
 
-Packet packet_create(const size_t ndims, const size_t* dims, const size_t* offset,
+Packet packet_create(const size_t ndims, const uint64_t* dims, const uint64_t* offset,
                      const size_t nbuffers);
 
 void packet_wait(Packet* packet);
